+++ src/main/java/com/zen/fogman/entity/the_man/TheManEntityHallucination.java
@@ -6,22 +6,10 @@ import net.minecraft.world.World;
 public class TheManEntityHallucination extends TheManEntity {
     public TheManEntityHallucination(EntityType<? extends TheManEntity> entityType, World world) {
         super(entityType, world);
-
-        this.startChase();
     }
 
     @Override
     public boolean isHallucination() {
         return true;
-    }
-
-    @Override
-    protected void mobTick() {
-        super.mobTick();
-        if (!this.getWorld().isClient()) {
-            if (this.isAlive()) {
-                this.setHealth(this.getHealth() - 4f);
-            }
-        }
     }
 }