+++ src/main/java/com/zen/fogman/item/ModItems.java
@@ -2,6 +2,7 @@ package com.zen.fogman.item;
 
 import com.zen.fogman.ManFromTheFog;
 import com.zen.fogman.entity.ModEntities;
+import com.zen.fogman.item.custom.ClawsItem;
 import com.zen.fogman.item.custom.TearOfTheManItem;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroupEntries;
@@ -9,12 +10,20 @@ import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroups;
 import net.minecraft.item.SpawnEggItem;
+import net.minecraft.item.ToolMaterials;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Rarity;
 
 public class ModItems {
+
+    public static final Item CLAWS = registerItem("mans_claws",new ClawsItem(
+            ToolMaterials.NETHERITE,
+            2,
+            1f,
+            new FabricItemSettings().maxCount(1).maxDamage(40).rarity(Rarity.EPIC).fireproof()
+    ));
 
     public static final Item TEAR_OF_THE_MAN = registerItem("tear_of_the_man",new TearOfTheManItem(
             new FabricItemSettings().maxCount(1).rarity(Rarity.EPIC)
@@ -39,8 +48,8 @@ public class ModItems {
         entries.add(THE_MAN_HALLUCINATION_SPAWN_EGG);
     }
 
-    private static void addItemsToFunctionalItemsGroup(FabricItemGroupEntries entries) {
-        // entries.add(TEAR_OF_THE_MAN);
+    private static void addItemsToCombatGroup(FabricItemGroupEntries entries) {
+        entries.add(CLAWS);
     }
 
     public static Item registerItem(String name, Item item) {
@@ -50,7 +59,7 @@ public class ModItems {
     public static void registerModItems() {
 
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.INGREDIENTS).register(ModItems::addItemsToIngredientItemsGroup);
-        ItemGroupEvents.modifyEntriesEvent(ItemGroups.FUNCTIONAL).register(ModItems::addItemsToFunctionalItemsGroup);
+        ItemGroupEvents.modifyEntriesEvent(ItemGroups.COMBAT).register(ModItems::addItemsToCombatGroup);
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.SPAWN_EGGS).register(ModItems::addItemsToSpawnEggsGroup);
     }
 }