+++ src/main/java/com/zen/fogman/item/custom/TearOfTheManItem.java
@@ -1,5 +1,6 @@
 package com.zen.fogman.item.custom;
 
+import com.zen.fogman.other.MathUtils;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffectInstance;
@@ -32,31 +33,10 @@ public class TearOfTheManItem extends Item {
     public void inventoryTick(ItemStack stack, World world, Entity entity, int slot, boolean selected) {
         if (!world.isClient()) {
             if (entity.isPlayer() && entity instanceof ServerPlayerEntity player) {
-
                 if (selected || player.getOffHandStack() == stack) {
-                    if (!hasEffects) {
-                        if (!player.hasStatusEffect(StatusEffects.NIGHT_VISION)) {
-                            player.addStatusEffect(new StatusEffectInstance(StatusEffects.NIGHT_VISION, StatusEffectInstance.INFINITE));
-                        }
-
-                        if (!player.hasStatusEffect(StatusEffects.SPEED)) {
-                            player.addStatusEffect(new StatusEffectInstance(StatusEffects.SPEED, StatusEffectInstance.INFINITE));
-                        }
-                        hasEffects = true;
-                    }
-                } else {
-                    if (hasEffects) {
-                        if (player.hasStatusEffect(StatusEffects.NIGHT_VISION)) {
-                            player.removeStatusEffect(StatusEffects.NIGHT_VISION);
-                        }
-
-                        if (player.hasStatusEffect(StatusEffects.SPEED)) {
-                            player.removeStatusEffect(StatusEffects.SPEED);
-                        }
-                        hasEffects = false;
-                    }
+                    player.addStatusEffect(new StatusEffectInstance(StatusEffects.NIGHT_VISION, MathUtils.secToTick(12)));
+                    player.addStatusEffect(new StatusEffectInstance(StatusEffects.SPEED, MathUtils.secToTick(12)));
                 }
-
             }
         }
         super.inventoryTick(stack, world, entity, slot, selected);